Reformatted Road Networks

+++ Data Generation/Reformat-RN.py
@@ -26,27 +26,21 @@ def pois():
             poi_id = int(float(row[0]))
             poi_dict = {}
             poi_dict["id"] = poi_id
-            poi_dict["category"] = int(row[1])
+            poi_dict["category"] = row[1]
             poi_dict["location"] = [float(row[2]), float(row[3])]
             poi_dict["keywords"] = []
-            poiKeysPath = "../Datasets/RoadNetworks/California/POI_keywords.csv"
-            if not os.path.exists(poiKeysPath):
-                with open("../Datasets/RoadNetworks/California/POI_keywords.csv", 'w') as catFile:
-                    writer = csv.writer(catFile, delimiter=',',)
-                    writer.writerow(["category_id", "category"])
-                    writer.writerow([0, poi_dict["category"]])
-                    categories += poi_dict["category"]
-                    poi_dict["category"] = 0
+            if poi_dict["category"] not in categories:
+                categories += [poi_dict["category"]]
+                poi_dict["category"] = len(categories) - 1
             else:
-                highestID = 0
-                with open("../Datasets/RoadNetworks/California/POI_keywords.csv", 'r') as catFile:
-                    r = csv.reader(csvfile, delimiter=',', quotechar='|')
-                    next(r)
-                    for row2 in reader:
-                        highestID = row2[0]
-
+                poi_dict["category"] = categories.index(poi_dict["category"])
             with open(f"../Datasets/RoadNetworks/California/POIs/{poi_id}.json", "w") as f:
                 json.dump(poi_dict, f, indent=4)
+    with open("../Datasets/RoadNetworks/California/POI_categories.csv", 'w', newline='') as catFile:
+        writer = csv.writer(catFile, delimiter=',',)
+        writer.writerow(["category_id", "category"])
+        for x in range(0, len(categories)):
+            writer.writerow([x, categories[x]])
 
 
 def loc():
@@ -68,24 +62,24 @@ def loc():
 
 def key():
     data = []
-    user_keyword_dict = {}
-    with open("../Datasets/SocialNetworks/Gowalla_old/gowalla_key.csv", 'r') as kfile:
+    poi_keyword_dict = {}
+    with open("../Datasets/RoadNetworks/California_old/california_poi_key.csv", 'r') as kfile:
         reader = csv.reader(kfile, delimiter=',', quotechar='|')
-        next(reader)
         for row in reader:
-            user_id = int(float(row[0]))
+            poi_id = int(float(row[0]))
             keyword_id = int(float(row[1]))
-            if user_id in user_keyword_dict.keys():
-                user_keyword_dict[user_id].append(keyword_id)
+            if poi_id in poi_keyword_dict.keys():
+                poi_keyword_dict[poi_id].append(keyword_id)
             else:
-                user_keyword_dict[user_id] = [keyword_id]
-        for item in user_keyword_dict:
-            with open(f"../Datasets/SocialNetworks/Gowalla/Users/{item}.json", "r") as f:
-                data = json.load(f)
-                data["keywords"] = user_keyword_dict[item]
-                f.close()
-            with open(f"../Datasets/SocialNetworks/Gowalla/Users/{item}.json", "w") as f:
-                json.dump(data, f, indent=4)
+                poi_keyword_dict[poi_id] = [keyword_id]
+        for item in poi_keyword_dict:
+            if os.path.exists(f"../Datasets/RoadNetworks/California/POIs/{item}.json"):
+                with open(f"../Datasets/RoadNetworks/California/POIs/{item}.json", "r") as f:
+                    data = json.load(f)
+                    data["keywords"] = poi_keyword_dict[item]
+                    f.close()
+                with open(f"../Datasets/RoadNetworks/California/POIs/{item}.json", "w") as f:
+                    json.dump(data, f, indent=4)
 
 
 def edge():
@@ -119,9 +113,9 @@ print("Creating nodes...")
 nodes()
 print(LINE_UP, "Adding node edges...")
 edge()
-'''print(LINE_UP, "Adding user locations...")
-loc()
+print(LINE_UP, "Adding pois...")
+pois()
 print(LINE_UP, "Adding keywords...       ")
 key()
-'''
+
 print("Complete!")
